requirements.txt
streamlit
pdfplumber
pillow
pytesseract
transformers
torch
groq
python-dotenv